requirement.txt
fastapi
uvicorn
pandas
numpy
pytesseract
opencv-python
pillow
python-multipart
matplotlib
scikit-learn
aiofiles
sqlalchemy
